backend/utils/argon2_utils.py
from argon2 import PasswordHasher
from argon2.exceptions import VerifyMismatchError

ph = PasswordHasher()

def hash_password(password):
    return ph.hash(password)

def verify_password(hashed_password, password):
    try:
        return ph.verify(hashed_password, password)
    except VerifyMismatchError:
        return False
