requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12.4,<4.0
djangorestframework-simplejwt>=4.8.0,<5.0
argon2-cffi>=20.1.0,<21.0
redis>=3.5.3,<4.0
requests>=2.25.1,<3.0
psycopg2-binary>=2.8.6,<3.0
